projects/shop/shop/requirements.txt
asgiref==3.8.1
django==5.2.3
djangorestframework==3.16.0
pip==24.0
sqlparse==0.5.3
